requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cloudinary==1.36.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers @ git+https://github.com/adamchainz/django-cors-headers.git@a4ab667a6e158e435287ad69914e9035265e841a
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
